+++ train.go
@@ -7,21 +7,18 @@ import (
 	"net/http"
 )
 
-// DetectRes .
-type DetectRes struct {
-	Count int            `json:"count"`
-	Faces []DetectedFace `json:"faces"`
+// TrainRes .
+type TrainRes struct {
+	FaceToken  string `json:"face_token"`
+	ImageToken string `json:"image_token"`
+	ImageURL   string `json:"image_url"`
 }
 
-// DetectedFace .
-type DetectedFace struct {
-	Rectangle Rectangle `json:"rectangle"`
-}
-
-// Detect faces within a given image.
-func (c *Client) Detect(image io.Reader) (*DetectRes, error) {
+// Train the model by uploading face image and tagging it with something unique to your app (face_id)
+func (c *Client) Train(image io.Reader, faceID string) (*TrainRes, error) {
 	var buf bytes.Buffer
 	w := multipart.NewWriter(&buf)
+	w.WriteField("face_id", faceID)
 	fw, err := w.CreateFormFile("image", "image.jpg")
 	if err != nil {
 		return nil, err
@@ -33,14 +30,14 @@ func (c *Client) Detect(image io.Reader) (*DetectRes, error) {
 		return nil, err
 	}
 
-	req, err := http.NewRequest("POST", c.baseURL+"/detect", &buf)
+	req, err := http.NewRequest("POST", c.baseURL+"/train", &buf)
 	if err != nil {
 		return nil, err
 	}
 
 	req.Header.Set("Content-Type", w.FormDataContentType())
 
-	res := DetectRes{}
+	res := TrainRes{}
 	if err := c.sendRequest(req, &res); err != nil {
 		return nil, err
 	}