+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+April 23, 2020
+
+- Train func
+- Helper func sendRequest
+
 April 22, 2020
 
 - NewClient, Detect